Added basic drawing functionality.
The drawing is implemented by creating a QPoint every time the mouse is
clicked and moved within the canvas. This point is then pushed into the
aptly named pointVector. The canvas is updated every time this happens,
and the contents of the vector are redrawn every frame.

A bug in this implementation is that rapid mouse movements leave gaps
imbetween the points. This can be fixed by drawing lines imbetween the
points, connecting them according to when the mouse was clicked.