Fixed bug which made different strokes have their endpoints connected.
This has been implemented via a new class, FlaggedQPoint. Its
constructors mirror those of QPoint, and the usage is mostly identical.
However, there is a public boolean flag, isConnected, which is meant to
signal whether or not the painter should draw a line between the current
FlaggedQPoint and the previous one within pointVector.