The decoded x coordinate printed by serialreceiver is not the correct x coordinate.
The code is currently printing p.x(), p.y(), isConnected on one line and
then the decoded x coordinate or cleared on the next line. The length of
the message sent is the correct size (34 bits) but I'm not sure whether
I decoded the message incorrectly or got the order of the bits mixed up
etc.